updated on 09/14/2017

+++ days/j05/ex04/ft_strncpy.c
@@ -6,21 +6,21 @@
 /*   By: pde-rent <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/09/05 04:47:57 by pde-rent          #+#    #+#             */
-/*   Updated: 2017/09/05 04:48:00 by pde-rent         ###   ########.fr       */
+/*   Updated: 2017/09/12 15:48:16 by pde-rent         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 char	*ft_strncpy(char *dest, char *src, unsigned int n)
 {
-	int i;
+	unsigned int i;
 
 	i = 0;
-	while (src[i] != '\0' && i < (int)n)
+	while (src[i] != '\0' && i < n)
 	{
 		dest[i] = src[i];
 		i++;
 	}
-	while (i < (int)n)
+	while (i < n)
 	{
 		dest[i] = '\0';
 		i++;
